202310/requirements.txt
PyYAML==6.0
accelerate
datasets==2.14.5
deepspeed
einops==0.6.1
evaluate==0.4.0
peft==0.5.0
protobuf==4.23.3
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
torch==2.0.1
transformers==4.34.0
fire==0.5.0
mpi4py==3.1.4
